:fire: Strip down groupbuild detail view
