+++ src/brouwers/groupbuilds/models.py
@@ -145,11 +145,6 @@ class GroupBuild(models.Model):
 
     num_participants.short_description = _("# participants")
 
-    def get_bbcode(self):
-        if self.pk:
-            return "[gb={0}][/gb]".format(self.pk)
-        return None
-
     def set_calendar_dimensions(self, start, end, num_months=6):
         num_months = float(num_months)
         # number of months
